changes in controller

+++ src/main/java/com/zaga/SnapValue/Repositories/UserRepository.java
@@ -1,12 +1,10 @@
 package com.zaga.SnapValue.Repositories;
 
-import java.math.BigInteger;
-
 import org.springframework.data.repository.CrudRepository;
 
 import com.zaga.SnapValue.model.UserModel;
 
-public interface UserRepository extends CrudRepository<UserModel, BigInteger> {
+public interface UserRepository extends CrudRepository<UserModel, Long> {
